Switch from a barrier between hpx_run instances to a condition.
We had originally designed HPX to sleep between hpx_run instances, and we
enforced this with a barrier. We don't really care if HPX sleeps though, it's
only a performance problem when its threads are running, not a correctness
issue.

Instead of enforcing a barrier, we just have a simple, traditional condition,
condition variable that indicates if we should be running. We don't want to
check this every time through _schedule() though, so we have a "hint" volatile
flag that we also set. When that flag transitions to SCHED_STOP we will acquire
the lock and check the condition.