Fix var-args issue with older compilers.
Now since we expect the second argument of marshalled actions to be a size_t, we
have to explicitly type-cast parameters that we provide for actions registered as
marshalled. This is annoying, and especially so for end-users, but I don't know
a good way other than reverting the second parameter back to an int.