Simplify the flush interface.
Now that we need to flush after every epoch it doesn't make sense to have such a
complicated interface for flushing. This does not fix ISIR, but it makes it a
bit easier to follow the code.

This also moves the flush out of `hpx_exit()` and into `_hpx_run()`. It isn't
safe to flush from the lightweight thread during stop because there is a chance
that we process our local stop parcel as work-first and that will cause us to
exit prior to flushing.