Fix the implementation of hpx_is_active().
The previous one checked the stopped flag, but that's not really adequate as
the active state might be stopped, but the local worker thread might not have
noticed yet. All sorts of craziness will ensue.

The only way to know if the current function was called from the outer SPMD
context is to check and see if the current thread pointer is set to non-NULL. A
conseqence of this is that the HPX worker threads will *always* be active, which
is exactly what we want.