Fix some blatent problems with profiline.
There was a large amount of trouble with the basic profiling implementation
related to memory allocation.

Also clobbered all instances of the volatile keywork in profiling. It's unclear
what the extensive use of volatile was for. If there is data that needs to be
sychronized than `sync` operations are necessary in addition to volatile where
appropriate.