Modify the C++ configuration to request "extended" C++11.
This means `-std=gnu++11` rather than `-std=c++11`. We need this for the same
reason that we need `-std=gnu99`---our preprocessor dependency on `, ##` for
`__VA_ARGS__` needs the extended preprocessor.

I also hoisted the C++11 checks into configure.ac, rather than leaving them
conditional in the libcuckoo and tbbmalloc m4 because we have tests that use
C++, and we'll be including the C++ bindings going forwards.