Switch to using exceptions to unwind hpx_thread_exit.
This replaces the previous unwind.h implementation with a C++ exception
implementation. The only benefit of the unwind.h implementation was that it was
C-only, however it didn't work on the arm platforms. After much debugging we
could not determine why it wouldn't work, so switched to the cross-platform
exception implementation.

We already required C++03 to build hpx examples, so this change doesn't have any
impact on the underlying tool set needed. It doesn't require C++11, nor does it
introduce overheads not present in the unwind.h implementation.