Rework pwc network command infrastructure now that they are encapsulated.
We used to use the HPX_ACTION infrastructure to register and call these
commands. This provided the potential for users to create their own PWC network
handlers, but this functionality was never used---nor was it obvious how to do
so if you wanted to.

The new technique just builds a static handler table for PWC-only use.