Update action registration functions and macros.
We used to explicitly take an action "handler" as part of the interface. In the
push towards making registration more generic we drop that requirement, and now
take the handler as part of the va_args, which are interpreted in a
type-specific way.

In the future this will allow us to register other kinds of actions, like C++
actions, where it is actually the class that defines a specific entry function,
rather than a generic entry function that forwards to a pointer.