Eliminate the ffi-specific cif field in the action table.
The vectored opencl actions already have an "environment" pointer as part of
this structure, and now that we've pushed all of the ffi framework inside of the
ffi.c action file, we can rely on that environment to be the ffi_cif pointer for
cif action. This eliminates a field in the action structure, which is important
since we allocate so many records in the actions table.