Changed initialization of constexpr static member to a proper form.
Previously, initializing the `arg_types_tpl` when compiling with `-O0` was
causing problems with uninitialized symbols.  It turns out that static consexpr
symbols do not get an actual address until they are defined outside of the
class they belong to.  This commit changes the previous fix to the proper
way of giving a static constexpr member an address useful for linking.