Updates to the C++ API.
The HPX headers were not designed to be included individually, but the C++ API
is doing that. I added __cplusplus checks to the HPX headers to avoid having to
put them in the hpx++ headers.

I also added a system.h file to wrap the hpx_bcast routine in a variadic
template instead of in a macro.