Remove the bsize from global_ptr<void>.
We need to be able to initialize global pointers to void directly from
hpx_addr_t without knowing anything about their block size, so we remove that
from the implementation. Any code that depends on bsize from a global pointer
will need to be strongly typed.

To deal with this we implement an explicit cast constructor to convert from
global_ptr<void> to typed pointers. The user obviously needs to know the bsize
to do this.