Add HPX++ wrappers for par_for.
There is the possibility that we could eitehr use std::function to pass a
lambda, or we could manually convert the lambda to a function/env type using
fancy metaprogramming. The first option has overhead that we don't currently
want to deal with, and the second case is complicated enough that we couldn't
figure it out.

Ultimately we want to be able to accept and apply lambda functions for par
for like other C++ frameworks (e.g., Kokkos). At this point however, we require
that the user explicitly deconstruct any lambdas that they'd like to use into a
properly-typed nocapture lambda and an explicit capture environment, since
that's how HPX++ works.