agas: fix writing into invalid local memory.
In AGAS, the root locality requests memory from the cyclic heap, and the
base pointer of this memory could be at an offset from the chunk base. This
offset is stored in the gva and is used during translation (gva->lva).

If the cyclic heap were to only manage our cyclic address space, we could
have set this offset to 0. BUT the memory we get back from the cyclic heap
is used as backing memory for the allocation on rank 0. So we just account
for this offset when allocating local memory on ranks != 0.