Remove an unused file from the pwc network.
Photon used to provide a fixed number of pwc/gwc slots to HPX, so we had some
buffering capabilities in order to shield the user from the limited
resources. Photon now provides its own internal buffering so the pwc/gwc
operations will not fail for resource reasons. We updated the rest of the
transport to work correctly with the new model, but left behind the pwc_buffer.c
file that implemented the buffering.

This file no longer has any external users (it's not even compiled) so I am
removing it to eliminate the chance of confusion. It can be resurrected from git
if necessary.