agas: Add AGAS-based dynamic load-balancing.
This commit adds dynamic, global, centralized load-balancing support to
HPX-5. It maintains a thread-private BST (Block Statistics Table) on
each locality which tracks GAS access to blocks. When agas_rebalance()
is called, these thread-private BSTs are aggregated into a per-node
BST which are then sent and aggregated at the locality that initiated
the rebalance operation.

Finally, the generated graph is fed to a partitioner library (presently
sequential METIS) which returns partitions for the blocks. The AGAS
rebalancer initiates "move" operations in parallel to rebalance the
GAS blocks.

TODO:
  1. Sanity check that the partitions look okay.
  2. Add support for task-scales or other partitioning library. Eventually
     we would want to use a distributed partitioner (perhaps ParMETIS?).
  3. Remove duplicate "vertices" from the aggregate global BSTs. We do this
     on a per-node basis, but if blocks were moved during the rebalance operation,
     then our graph will have duplicate vertices. For now, we assume that there
     are no concurrent move operations when AGAS is rebalancing.