Implement the rebalancer "move" optimization.
After partitioning the block graph, we issue move requests to all of the
blocks in parallel. This is inefficient because if a block is not supposed
to be moved, we still send a move request and we wouldn't know if it needs
to be moved or not until we go to the block's owner.

This change adds an owner map to the AGAS graph structure. When issuing
move requests in parallel, we look at the owner map to determine if a block
needs to be moved.