Add parcel compression.
This change adds a "compressed" network that compresses parcels on
the fly. The lz4 compression scheme is used because it provides
fast compression and decompression rates. Presently, we do not do
in-place compression and as such the network incurs extra copies
on the send and receive sides. The "--hpx-parcel-compression" option
can be used to enable this feature. Only actions tagged with the
HPX_COMPRESSED attribute will be compressed.