Fix invalid free during AGAS move.
We account for moved blocks during agas-free but simply by comparing
the current owner with the home. This leaves a lot of cases where we
could potentially leak local memory. In particular, we always have
to free for the following two cases:
 1.  If it is a single block, or
 2.  If we are moving from a non-home owner to another owner.

There are cases we don't handle still, for instance:
 move from home -> owner -> home.