Correctly sycnhronize access to the work_id variable.
This worker variable is read during stealing and is thus a potential data
race. We don't really care about the race---we don't care what value we see
there as long as it is either 0 or 1---but the C/C++ standard says that we could
see *anything* if it's a race (in fact the entire program is undefined).

This commit updates all accesses to this variable to use the sync interface with
SYNC_RELAXED semantics, which has no relevant overhead but would preclude
certain kinds of optimizations (e.g., value speculation) that might be
dangerous. None of these would actually be done so this is free.x