Fixed address addition to correctly handle negative offsets.
The problem was when a large negative offset was added to an address.
First of all, some of the intermediate results could be negative, making it
necessary to use signed integers.  Second, negative offsets require "wrapping
around" of parts of an address (e.g., if `phase` is 0, subtracting one requires
phase of `bsize - 1`).