Restructure the scheduling code.
1) Split the unified transfer function into a blocking and fast non-blocking
version. All of the transfers will start as fast non-blocking transfers. If this
fails to find work, it will transfer back to the system thread, running the
transfer continuation quickly, and then go through a more expensive blocking
epoch. This change eliminates the "block" paramemter to scheduler_suspend---all
uses of scheduler_suspend are now considered to be non-blocking (with respect to
the passed continuation).

2) Simplify some of the control flow.

3) Restructure the file to group all of the checkpoint continuations in the same
place.