Restructure the scheduler so that the main thread is not a worker thread.
For a long time the scheduler has subsumed the main program thread. This has
worked well up to this point but complicates a number of issues.

1) The code paths into and out of hpx_run are quite complicated.

2) The hpx_run infrastructure is having strange problems with repeated
high-frequency hpx_run instantiations.

3) The APEX infrastructure is going to want to start shutting down specific
cores.

4) We want to experiment with various asynchronous activities like network
progress, timer LCOs, and GPU polling.

Given these challenges this commit separate the infrastructure into two pieces,
the primary thread and the scheduler threads. A side effect of this commit is
that it simplifies the code and provides better abstraction through some header
modifications.