Invert the heap layout for PGAS.
This patch switches to using the top half of the heap segment for cyclic
allocations and the bottom half of the heap segment for global accesses. This
will attempt to address the problems with the hpx-143 fix for large cyclic
allocations by allowing the PGAS infrastructure to reserve the `0` as a global
variable at rank 0.