Update hpx-143 fix.
The hpx-143 bug is caused when the first global allocation is "huge" (or has a
"huge" alignment) from the perspective of jemalloc. In that context, jemalloc
won't use a chunk header but will instead use an out-of-place header, and the
0th offset of the 0th chunk of the 0th rank will be returned, resulting in
HPX_NULL as a valid address. The old fix was doing an initial small cyclic alloc
to deal with this, however now that we have inverted the heap, that 0th offset
corresponds to a global alloc.

This fix removes the cyclic allocation and converts it to a small global
allocation at rank 0. This should improve startup time, reduce memory overheads,
and eliminate the fragmentation caused by the old solution.