Handle the ssync network_send continuation in pwc.
This implementation co-opts the `next` field from the primary parcel to store
the ssync parcel pointer. All parcels that are given to the network will have
the `handle_parcel_delete` pwc network command run on them sooner or later after
the underlying network operation happens (either directly though a local
completion command, or indirectly through a rendezvous send completion), so
during this command we just check and launch the ssync parcel if there is one
stored in `next`.