Update the `HPX_GAS_BLOCK_BYTES_MAX` to be a GAS-specific value, and extend the GAS interface to support `size_t` block sizes.
The HPX_GAS_BLOCK_BYTES_MAX can be used by the application to decompose data
structures in an application specific way, but it depends on the GAS that has
been selected dynamically. For example, our PGAS implementation limits blocks to
2^32 bytes, but the SMP implementation has no obvious bound.

In order to use this effectively, the block-size interface needs to be updated
everywhere to use `size_t` block sizes as well. Specific implementations that
need smaller block sizes can test dynamically and then do a narrowing cast.