Remove special-case entry point for interrupts.
We used to have a special-case handling mechanism for threads tagged as
"interrupts" by the user. This would launch the interrupt using the parent
thread's stack without calling transfer.

These two operations are both relatively low overhead in HPX, and its not
obvious that there is really any benefit, particularly given the complexity of
the code. It also complicates the code path for spawn(), which we want to
avoid.

HPX-950 #comment