Add an initial implementation of hpx_gas_{set,clear}_affinity.
This uses a static libcuckoo hash table to store affinities. It promotes the
libcuckoo contrib to required, which transitively makes C++11 support
required. In the future it might make sense to add a configure-time option that
can enable and disable gas affinity in order to remediate this increased
system depenendecy. At the same time, all of the systems that we work with
_have_ C++11 support, so it's not a huge concern.