Add an implementation of affinity using userspace-rcu.
This implementation leverages the fast read-only capabilities of the userspace
rcu library for an initial implementation of affinity. There are some
optimizations to be made, e.g., Node freelisting or deferred deletion in order
to avoid synchronization during insert and remove, but it is much more efficient
than libcuckoo for small scales.