Stop relying on global LIBHPX_COMM and initialized flags for MPI.
We would like to be more friendly to applications that want to mix HPX and
MPI. One part of that is allowing the user to MPI_Init before calling hpx_init,
but our previous "is MPI initialized" testing didn't quite work.

It would be much easier and cleaner to just use the bootstrap and/or network
object to track the communicator and initialized flag. This commit does exactly
that by extending some of the member function interfaces that were previously
static (and relied on the global data) to be instance members.

The only complication is photon which wants to use the bootstrap
allreduce/barrier in a static way, so we needed to include a simple hack to
target the global here->boot network in that case.