agas: Fix root block at the allocating locality for user-defined data distributions.
In general, we do not support moving the root block of an allocation. This is true of
local array allocations, cyclic allocations and also user-defined allocations. This
restriction can be lifted by extending the block structure. For now, we simply ignore
the mapping for block 0 when performing a user-defined allocation.