Introduce a type for global pointer difference and offsets (hpx_gas_ptrdiff_t).
The HPX global address space is fixed to 64 bits, but in the future we might
want it to be architecture-specific, i.e. 32-bits on 32-bit platforms. In this
commit, we abstract the global pointer difference type (hpx_gas_ptrdiff_t) which
should be used for global pointer arithmetic and array indexing.

On non 64-bit platforms, users should take into account sign extension rules
for implicit conversions to this type, and use explicitly conversion/casting
whenever possible.