Fix a really sneaky race under debug executions.
The LCO is designed to check to make sure that there are no "nested monitor"
lock calls for a thread by setting a flag in the thread's control block to
indicate that it owns a lock. Given that this data is thread-local, I was
manipulating it without a lock.

This behavior wasn't safe because there is a tiny window of time inside the
transfer continuation, right after the lock is unlocked, where the thread can be
resumed be a different system thread. This is fine---the thread is no longer
running on the original system thread---however modifying the thread's control
block during this window is unsafe as it will be concurrently updated when it is
resumed.

The fix to this is simple, just protect the update with the critical region.