Stop requesting a specific address for the PGAS global heap segment.
The PGAS heap segment needs a very large alignment so that we can mask out the
offset from a local virtual address. Historically, we enforced this by asking
for a nicely aligned address during mmap().

More recently we have implemented an "aligned" mmap() wrapper that will
overallocate pages and find the right aligned address and trim the excess. It is
no longer important to ask for a specific address during heap allocation. This
provides more flexibility about where in the address space the system can
satisfy this allocation from.