Update stack freelisting to avoid undefined behavior.
We always freelisted threads through their "next" pointer. In the C++ conversion
we started explicitly calling the thread destructor `~Thread()` before
freelisting. Once this happens it is no longer safe to inspect or modify the
threads' fields, e.g., setting the next pointer.

Normally this wasn't an issue, but with aggressive link-time-optimization this
triggered a runtime error. I workaround this issue by adding an explicit
freelist node type, and using placement new to allocate one on top of the thread
buffer.