Workaround some icc issues.
The intel compiler has an ICE when explicitly initializing fields that are
arrays of std::atomic. This occurs in the process Bitmap, so we just leave that
field implicitly initialized.

The intel compiler has issues with the stdatomic.h header that I was including
in some unit tests post-libsync removal, so I switched those tests to compile
with C++ and to use C++ atomics.