Fix a bug related to pwc eager buffer reference counting during shutdown.
The pwc eager buffers are reference counted. When the number of bytes remaining
in the buffer reaches 0 it is automatically freed. It is normally the case that
the "reload" mechanism has replaced the current Peer-referenced buffer before
the reference count reaches 0 because the local scheduler hasn't yet processed
all of the parcels that were sent. At shutdown the Peer deletes the current
buffer.

It turns out that there is an extreme corner case race where the current buffer
has had its last byte reclaimed (and thus has been deleted), but no reload
command has yet run. In this context the current recv_ pointer in the Peer is
pointing at a buffer that was already deleted, so deleting it at shutdown is a
double-delete and triggers allocator issues.

The fix for this is to check and see if the current recv_ buffer has already
been reclaimed at shutdown before trying to delete it.