Add a libhpx interface to get a parcel source locality.
Inside of the libhpx implementation the parcel structure tracks the sender
source id. This isn't exposed as part of the HPX API because we don't want to
require that all implementations have this capability, however when a client is
depending on the libhpx.h header we can provide this for them.