Move to heavyweight network progress.
This shifts from trying to use a lightweight thread to progress the network to
using the scheduler pthread to progress the network. The thread is set up with a
progress timer that wakes it up periodically, and any of the scheduler threads
can kick the thread to get it to run an on-demand progress.