Use a priority queue for the global queue.
This introduces a PriorityQueue interface and uses a public version of the
linden/jonsson queue from http://user.it.uu.se/~jonli208/priorityqueue/ as a
first implementation. The LJQ was updated to store pointer values and to sort by
MAX rather than by MIN in order to meet our expected semantics.