Expose basic condition synchronization for threads through libhpx.h.
HPX models synchronization through LCOs but they are pretty restricted to raw
dataflow style operations. There are more complex shared-memory condition
synchronization problems that are difficult to encode as pure dataflow, like
producer-consumer queues. It is possible to implement these as dataflow, but
until we have a good API for doing so we expose the underlying libhpx
condition-synchronization infrastructure for lightweight threading.