Mark that we're holding a lock inside the cond variable interface.
The libhpx scheduler knows that it shouldn't do anything cray (like context
switching) when a lightweight thread is holding a lock related to a condition
variable (i.e., is "in" an LCO in HPX parlance). This patch makes sure that the
libhpx_cond API provides the same information.

This resolves a bug where, when the scheduler has been flooded with parcels, it
was trying to do a work-first context switch inside of the libhpx_cond_broadcast
function, resulting in a deadlock.